+++ scripts/core/ui/components/Form/SelectMetaTermsInput/SelectFieldPopup.jsx
@@ -337,7 +337,8 @@ export class SelectFieldPopup extends React.Component {
                                     </div>
                                 </button>
                             </div>
-                        ) || (
+                        )
+                        || (
                             <SearchBar
                                 onSearch={this.filterSearchResults}
                                 minLength={1}
